refactor to mixins and add type checks Re #151
